BF: Set phase encode to 'ROW'.
Nifti and dicom represent x/y differently-- dicom uses a ROW,COL (Y,X)
convention, and nifti uses the more familiar X,Y convention. So, a dicom
COL corresponds to the first dim (X) in the nifti. In the older version
of our dicom-to-nifti conversion, we explicitly flipped the x and y dims
in the data (and correspondingly adjusted the affines) so the data would
look the same as they look on the scanner (A/P is up/down). So, in that
framework COL in dicom == Y in nifti. When changing over to dcmstack, we
decided not to flip X/Y, but forgot to change the phase encode
calculation to match. - RFD