Remove recon_mode_flag check
This dicom tag (defined in line 29) is a flag for an image space filter applied to the dicoms, we used to use this filter for the multi-coil SPGR scans, so this line has been working for those scans. But actually this flag itself is not an indicator of multi-coil scan. Now we removed this filter in those multi-coil scans, so the recon_mode_flag is now 0, and this caused the nifti conversion to fail for those scans. Removing the check should solve the problem.